Added work experiences, project dates, images, and changed the website page accessibility
Added code to support buttons and sections for work experiences, largely based on the code for articles. Added dates to all projects and work experiences. Since what I'm displaying on the site is simpler than what is in the template, I removed the navbar from each page and simplified website navigation: pages for projects and work experiences are only accessible through the URL or by clicking on their corresponding links on the homepage, and the back button returns the user from a project or work experience page back to the home page. Added some images, such as the portrait on the home page and the website icon.